\name{DebugExample-package}
\alias{DebugExample-package}
\alias{DebugExample}
\docType{package}
\title{
\packageTitle{DebugExample}
}
\description{
\packageDescription{DebugExample}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{DebugExample}

Maintainer: \packageMaintainer{DebugExample}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = DebugExample)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{DebugExample}}
%\section{Documentation Index}{\packageIndices{DebugExample}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
